requirements.txt
# Django and REST Framework
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0

# Environment Variables
python-dotenv>=0.19.0,<1.0.0

# Firebase
firebase-admin>=5.0.0,<6.0.0

# Django Channels
channels>=3.0.0,<4.0.0
asgiref>=3.4.0,<4.0.0

# HTTP Libraries
requests>=2.25.0,<3.0.0

# Machine Learning Libraries
scikit-learn>=1.0.0,<2.0.0
pandas>=1.3.0,<2.0.0
numpy>=1.20.0,<2.0.0
joblib>=1.0.0,<2.0.0

# Authentication
PyJWT>=2.1.0,<3.0.0

# WebSocket support for Django Channels
daphne>=3.0.0,<4.0.0

# CORS Headers
django-cors-headers>=3.10.0,<4.0.0

# PDF Generation
reportlab>=3.6.0,<4.0.0
weasyprint>=53.0.0,<54.0.0

# Image Processing
Pillow>=8.0.0,<10.0.0

# Background Tasks
APScheduler>=3.10.1,<4.0.0
